Use more appropriate HTML elements for annotation card sections
`<section>` is really intended for a stand-alone section of content on
the page which _should generally have a heading and would make sense in
an outline view of the page [1]_. It isn't appropriate for the
quote/body/tags in this context.

See https://developer.mozilla.org/en/docs/Web/HTML/Element/section#Usage_notes

Additionally removing the title for the annotation body makes scanning
the content easier.

[1] See https://gsnedders.html5.org/outliner/