Add group creator to members in service
Move the code that adds a group's creator to its list of members from
the group model to the group service.

This means that other code that creates Group ORM objects directly (such
as another service method) can create groups whose creators are not in
their lists of members (as is desired for "publisher groups"). Meanwhile
code that does still want to create a group with its creator as a member
can do so via the service.