Upgrade pyramid_authsanity to 1.0.0
[Our fix](https://github.com/usingnamespace/pyramid_authsanity/pull/7)
for an esoteric IE bug (where we'd set a correctly signed cookie instead
of removing a cookie) is in the latest release of this package.

There are a couple of small changes to the generated requirements.txt
that aren't directly related to this upgrade, but they all look harmless
to me and are mostly the result of us acquiring direct dependencies on
packages that were previously transitive dependencies (such as
transaction, venusian and zope.interface).