Append UriCombinedWildcardFilter in /api/search
This will allow searching for uri's containing wildcards via the
wildcard_uri parameter.

Add the wildcard_uri parameter to the schema and raise a validation
error if the uri value is invalid (aka contains wildcards in the
domain of the uri).

Update the docs to reflect the additional wildcard_uri parameter.