Add UriCombinedWildcardFilter
Searching for a partial url is a common use case. It is particularly
useful for publishers who want to know who has been making annotations
on all their articles. It is also useful for users who simply can't
remember the entire url of an article they have annotated.

Add UriCombinedWildcardFilter has two configurations:
 1. seperate_keys=True
    Wildcards are only present in wildcard_uri param values and exact
    matches are present in uri/url param values.
 1. seperate_keys=False
    Wildcards and exact matches are present in uri/url param values.

UriCombinedWildcardFilter igores all wildcard uri's that have wildcards
in the domain.

Remove UriFilter from default modifier list in Search class since
the UriCombinedWildcardFilter will need to replace it in certain
cases and there is no way to selectively remove modifiers. The
current design pattern is if the modifier is used by all searches,
then it is in the default list, since it will not be, drop it from
the default list.

Add both UriFilter and UriCombinedWildcardFilter's to the search __init__
module as they will be appended to the Search class modifier list as needed.

Modify all existing code that runs Search to append the UriFilter to the list
of modifiers before running a search.