Expose OAuth client ID and feature flag to client.
Expose the embedded Hypothesis client's OAuth client ID to via
configuration in "app.html". This config is read from the
CLIENT_OAUTH_ID env var, not to be confused with the CLIENT_ID env var
used for the "iss" field of the legacy JWT tokens fetched via `GET
/api/token`.

In addition, add a separately toggle-able feature flag which is exposed
to the client via the "oauthEnabled" config setting. Having a separate
flag enables toggling it without changing the service config, as well as
enabling the page embedding the client or bookmarklet to control use of
this feature.