Add feature flag to append cachebuster to client URL
Add a feature flag which causes a cachebuster to be appended to the
URL of the client's boot script when serving the embed.

This provides a way to bypass the CDN cache on a per-cohort basis (eg.
temporarily enable for a dev after a client release for testing), or on
a per-request basis (eg. create a bookmarklet which serves
'.../embed.js?__feature__[embed_cachebuster]' and sets the sidebar app
URL to '.../app.html?__feature__[embed_cachebuster]').