Add AuthClient grant_type and response_type columns
In order to support the OAuth authorization code flow that the client is
intending to use, we need to be able to distinguish between clients that
are allowed to use this flow and those that aren't.

Adding grant_type and response_type columns to the AuthClient table
allows us to specify for a given client which response_type parameter is
valid for the authorization endpoint, and which grant_type parameter is
valid for the access token endpoint.

For both grant type and response type we enumerate all valid types for
each field from RFC6749 (The OAuth 2.0 Authorization Framework) and
RFC7523 (JSON Web Token (JWT) Profile for OAuth 2.0 Client
Authentication and Authorization Grants), even though we will not
initially support all the valid permutations. This is simply because
changing enum types in Postgres can be a bit tedious, so if we can avoid
doing that as and when we add support for (e.g.) implicit or resource
owner credentials (AKA "password") grants, we'll save ourselves some
pain.