Replace monkeypatch with explicit constructor parameter
The existing way of patching WebApplicationServer inside
OAuthAuthorizeController worked, mostly, but there was an implicit
dependency on the `oauth` fixture from the `controller` fixture, which
was easy to miss.

This commit doesn't change much about the details of the tests, but it
adds a test seam in the form of a constructor parameter to
OAuthAuthorizeController which can be used to pass in an alternative
OAuth Server profile class.