Add support for "trusted" OAuth clients
Certain clients are trusted, by which we mean that we know that they
can be trusted with an access token for a user. That means that we don't
ask the user to explicitly authorize trusted clients, and instead skip
straight to an authorized response (i.e. a redirect with an authz code
or in future a postMessage with same).