Use a wrapper for `request.client` in OAuth validator
Once the client is loaded from the database oauthlib expects
`request.client.client_id` to return the client id. Our `AuthClient`
models does not respond to `client_id`, so the cleanest solution is to
introduce a new wrapper class that responds to `client_id`, but also
allows access to the `AuthClient` instance itself.