Make cohort labels clickable in feature flag admin
In the current feature flag admin, there are labelled checkboxes to
enable individual features for each cohort of users we create. The
labels have a `for` attribute, but as they refer to the `name`
attributes of their respective checkboxes, rather than their `id`
attributes, clicking the cohort name doesn't toggle the checkbox (at
least not according to the spec, or testing it in Safari, Firefox and
Chrome).

Adding an `id` attribute identical to the `name` attribute fixes this
problem, giving our admins a slightly easier experience trying to click
things.