Check authclient grant type in user creation API
We currently allow any confidential auth client to create users in their
authority. This is a bad idea.

I've opted for checking that the configured grant type is
`client_credentials` and only then allowing these auth clients to create
users.
This works for now as we don't use the `client_credentials` grant type,
but once we do support it the way it is meant to be used by the spec, we
should rethink this and maybe add a boolean field for this.