Document the group-leaving endpoint
I'm not sure whether there's a good way to explicitly document that the
success response will be empty, but a human reader can at least infer
that from the fact that it's a `204 No Content` response.

While the endpoint is currently only applicable for leaving a group,
rather than removing anyone else, I've written the documentation (and
named the endpoint) in such a way that we can expand this to removing
other users and only have to change the validation rules on the user
parameter.