Update token userids rather than deleting them when renaming a user
As noted in [1], deleting rather than updating tokens when renaming a
user will break any applications which have been configured with a
developer token, which may be unexpected/annoying.

Unlike auth tickets, where the userid is part of the cookie value stored
in the user's browser, the userid for OAuth tokens only exists in h's DB
so we can just update it.

[1] https://github.com/hypothesis/h/pull/4631#pullrequestreview-57693684