Extract API config into its own module
This functionality is now shared between several API view modules, so it
makes sense to pull out the shared functionality into its own module
rather than have all the non-root view modules import it along with the
rest of the root.

This is a first step towards encapsulating some of this functionality.
For example, the `add_api_view` function never gets used directly,
except by the `api_config` function and the module tests, so that could
potentially be made private.