Separate link_name and route_name in link tests
These appear to be two different concepts in the API (link_name being a
dotted "path" to help construct the tree of API routes at /api/, and
route_name being the key for Pyramid's internal route naming). Having
them use the same value in the tests is a little confusing, and can
cause bugs to slip through.