Extract route templating into its own class (#4639)
The main reason for doing this is so that we can support additional
path parameters, which currently result in a `KeyError` in the view.
This also has the benefit of making the view logic cleaner, and the
intent of the parameter substitution code clearer.

In doing this, the value of the URL-safe parameter has also changed to
something less likely to conflict with different parameter names. The
pattern of `_id_` can cause incorrect results if the parameter names
are, say, `id` and `group_id`.

We've decided to pass the parameter names into the templater explicitly
rather than have the templater work them out dynamically, because the
best solution we could come up with for that relied on the `KeyError`
exception (thrown from within Pyramid's routing code) for flow control.
In the future, if we could find a neater way to introspect Pyramid's
routes and discover the parameters for ourselves, we could remove the
explicit parameter.

There are comments explaining the need for explicit parameters, both in
the routes file and in the view code, so if someone needs to introduce a
new route parameter they should hopefully either see a comment before
the functional tests break, or should find a comment in the first place
they look to diagnose the breakage.