Use the new URL template generator for links (#4642)
Now that we have a class to handle template generation, we can use it
for the user stream URL instead of substituting manually.

I haven't replaced the group-leaving URL here, even though it works in a
similar way, because there is already a separate pull request to remove
this view entirely, so there's no need to change it here. Also, it would
take extra work in the class to convert the `pubid` route parameter into
the `id` template parameter.