Validate the syntax of DOI values
Validate the syntax of DOI values before creating document equivalence
claim entries in the DB which are labeled as being of type `dc-doi` or
`highwire-doi`.

Not all Dublin Core identifier field values (from `<meta
name="dc.identifier">` tags in HTML documents) are DOIs.

Additionally, the identifier value is only intended to be "unambiguous
within a given context" [1] which I take to imply that it is not
necessarily globally unique. If the value is a DOI then it is however.

[1] http://dublincore.org/documents/dcmi-terms/#terms-identifier