Support DOI URLs
Examining existing DOI metadata stored in the h prod DB shows that
"dc.identifier" and "citation_doi" fields are often used to store DOIs
as URLs rather than just identifiers.

Ideally we would normalize the different representations to the same
form, but for now we just continue to accept them as-is.