Work around limitations of `window.postMessage` in IE 11
In IE 11, postMessage support is limited to frames embedded within a
page and doesn't work between a popup window and its opener. This means
that OAuth clients wanting to support this browser will need to use a
redirect rather than the "web_message" response mode.

For the specific use case of the Hypothesis client, we can work around
the problem by manually dispatching the "message" event, in the target
window, that a `postMessage` call usually results in. This only works
because the sidebar's "app.html" and the auth popup are served from the
same domain.

IE 11 usage is currently around ~1% of client views, so it seems likely
that we might just be able to drop support a few months from now.