Only preload users when we actually render annotations
When the list of annotation ids for which we want to preload the users
is empty, then SQLAlchemy runs a query that looks like:
`SELECT annotation.userid AS annotation_userid FROM annotation WHERE
annotation.id != annotation.id`.
We need to ensure that we actually have been given any annotation ids
before we try to warm up the user cache.