Split up `AnnotationUserInfoFormatter`
We will also include the `user_info` object in the API profile endpoint
which does not know anything about annotations. So this is splitting up
the responsibility and putting it into a `user_info` function in the
`h.session` module. The `AnnotationUserInfo` formatter will then use
that function to generate the object.