Remove the legacy homepage
The legacy homepage has been replaced on the production instance of "h"
by the new property at https://web.hypothes.is. Therefore we can remove
the assets from the "h" service.

When a user visits the index route, they will either be redirected to
the preconfigured homepage URL or to the `/search` route if no redirect
is configured.