Simplify the logic in the login form's Cancel button and add tests
Going "Back" is not always going to be the correct way to "Cancel" the
form. In the context of the only use case we have at the moment, the
login popup, closing the window is always the way to cancel the form.

Unlike the Cancel button in the "Allow access?" screen, this does not
post an `authorization_required` message back to the client. It will be
the client's responsibility instead to watch for the window being
closed.