Add CLIENT_RPC_ALLOWED_ORIGINS setting
Add a new environment variable CLIENT_RPC_ALLOWED_ORIGINS. All h does is
read in this environment variable and pass it to the client as the
rpcAllowedOrigins client config setting.

This is for a new client feature that the client (sidebar app) responds
to JSON-RPC requests over postMessage from frames on other origins. The
origins whitelisted in the CLIENT_RPC_ALLOWED_ORIGINS settings are the
origins that the client will respond to such requests from.