Add OAuth to authorization docs
Rewrite the authorization docs for the API, moving them from the API
reference to Sphinx and describing how to a) use OAuth to obtain access
tokens and b) obtain client credentials from the OAuth control panel.

Moving the auth docs from ReDoc to Sphinx enables the "Authorization"
section to be presented in a logical order in the table of contents, as
the first subheading under "Using the Hypothesis API". It is also
enables easier interlinking with other content.