First stab at an API-based top-level importer
This is the barest minimum I could think of to get something that looks
API-compatible. It's not going to work yet, at least until we can pass
in an auth token and a group ID, and it completely ignores the problems
of ordering annotations into an order that respects replies, and reply
ID lookup, but it's a start.