Add eLife annotation timestamps data migration
We're going to run an API script to import 1000+ Disqus comment from
eLife into page note annotations using the standard annotation create
API. After being created in this way the annotations will have the wrong
timestamps - they'll be the times at which the annotations were created
in h, not the times of the comments exported from Disqus. This migration
sets the timestamps of these imported annotations to their correct
Disqus-exported timestamps.