Index authority of annotation creator as a separate field
This makes it possible to filter the annotations that a search returns
by authority. This can be used in order to eg. exclude annotations from
third party accounts from the public Hypothesis activity pages.

Unlike the current indexing for the "user" field, the preprocessing is
done in Python in order to be able to reuse the existing `split_user`
function.