Add a test for authority indexing that actually uses ElasticSearch
Previous tests for `h.search` always mocked ElasticSearch. However given
that the ES API is complex (both in params and response types) and that
we'll want to be able to verify everything is working after we upgrade
in future, it is useful to have tests that really do use ElasticSearch.