Add eLife test annotation timestamps data migration
eLife have run an API script to import 1000+ Disqus comment as page note
annotations in their test group, using the standard annotation create
API. After being created in this way the annotations have the wrong
timestamps - they'll be the times at which the annotations were created
in h, not the times at which the comments were originally created in
Disqus.  This migration sets the timestamps of these imported
test annotations to their correct Disqus-exported timestamps.