Add models.Annotation.authority property
Add a computed property `authority` to the annotation model.

This returns the authority (e.g. hypothes.is or elifesciences.org) of
the user who created the annotation and the group that the annotation
belongs to.

This will enable code in a future commit to decide whether an annotation
is a first- or third-party annotation given just the `Annotation`
object, without having to query for the annotation's user or group in
the DB.