Extract user deletion logic into a service
Follow the pattern of keeping non-trivial logic involving models out of
views by extracting it into a helper service.

This will also make it easier to reuse this functionality from other
views in future, such as an API for deleting users or a self-service
view for a user to delete their own account. The user deletion CLI
command was already inappropriately using the helper function from
`h.views.admin_users`.