Log a warning if a Pyramid request ends with uncommitted DB session changes
The previous `session.dirty` check only tested for _changes_ to existing
DB objects which have not been flushed. It did not check for added or
deleted model objects which had not yet been flushed to the DB, or for
changes which have been flushed inside a transaction but not yet _committed_.

When debugging https://github.com/hypothesis/h/issues/4704 we want to
specifically catch additions (of annotations) to the DB which have been
flushed (by `storage.create_annotation`) but not yet committed.