Filter out error logging from `ws4py`
In theory, this is useful logging to have. In practice, this produces a
large volume of errors of the form:

    Error when terminating the connection: [Errno 32] Broken pipe

Our best theory about these errors is that they come from
over-enthusiastic AWS connection termination. Either way, it's not
useful to us to have these messages, and even when we filter them out in
Sentry, they still have a nasty habit of overflowing our usage quotas.

It's worth noting that this is a slightly blunt tool with which to deal
with this problem, in that it will make it harder to see any errors
coming from the `ws4py` library without trawling through log files. The
other errors that have been known to occur in this library are:

    [Errno 9] File descriptor was closed in another greenlet
    [Errno 104] Connection reset by peer

I'm not sure either of these errors give us a whole lot of actionable
information either, at least at the moment.

The `logging` library does provide us with some more fine-grained
filters we could use, but that would require setting up the extra
code structures to set these filters up on the logging objects. This
would also come with the cost of having our logging filters set up in
config *and* code, rather than purely in configuration files.