Fix double-reporting of DB session warning to Sentry
We use Sentry's integration with the "logging" package to log all
messages >= WARNING to Sentry. Therefore by using `log.warn` and also
explicitly calling `request.sentry.captureMessage` we were capturing two
reports for the same event.

Since the change list was included in the format argument to `log.warn`
rather than being passed as a separate argument, Sentry was unable to
correctly group the reports.

Fix this by removing the `captureMessage` call and passing the changes
list as a separate argument to `log.warn` so that it appears in the log
output without affecting Sentry's ability to group reports.