Add a note about what happens if a subscriber throws an exception
It isn't clearly specified in the Pyramid documentation, but the
behavior is that if one event subscriber throws an exception, this can
prevent other subscribers to the same event from running.

This doesn't look to be easy to change, so at least flag it up as a
hazard for others to be aware of.