Allow logged-out users to view group pages for world-readable groups
Consolidate the logic for handling group read pages for logged-in and
logged-out users in `h.views.activity.GroupSearchController`, and let
logged-out users view world-readable groups.

With this change, it is now possible for a logged-out user to see the
activity page for a world-readable group such as "Public" or in future
open groups, eg. at https://hypothes.is/groups/__world__