Rename "publisher" to "open" in some tests
This is one of several commits renaming "publisher group" to "open group"
throughout h.

It turns out that we we originally called "publisher groups" when we developed
them for the eLife project (groups that are readable by anyone, writeable by
anyone in the same authority, and can't be joined) are the same kind of groups
as what we now want for the new "open groups" feature, so we're renaming the
existing "publisher groups" to "open groups" throughout.

The only difference is that groups like eLife's belong to a third-party
authority (they're "third-party open groups") whereas the new open groups we're
working on are going to be first-party.