Replace type_ param of create() with constant
The param is never used anymore so can be replaced with a constant. The
constant will itself be factored out later.

This removes create()s ability to create "open"-type groups or its need
to raise if type_ is invalid, so the code and tests for those behaviours
are removed too.