Replace FakeGroupService with autospecced mock
Replace a "manual" FakeGroupService object in some tests with an
autospecced mock object.

If GroupService changes (for example: adding, removing or renaming
methods of parameters) this mock will automatically update.

Additionally, it's not necessary to write a fake create() method that
keeps a list of all the arguments it's called with because mocks already
do this with their call_args_list.

The fake member_join() and member_leave() methods weren't used.