Clarify that userid of the creator is published
Clarify that it's the userid of the group's creator that's published
when the 'group-join' event is published when creating a group, by
publishing group.creator.userid instead of the local variable userid.

This will also make later refactoring easier as the publish line needs
only group and not the local userid variable.

Also fix a test: this group service test was expecting 'theresa' (the
username) to be published but in all cases what the real code actually
publishes is 'acct:theresa@example.com' (the userid) (see
h/views/groups.py and also h/cli/commands/groups.py, and note that the
local variable was called userid not username). Fix the test to expect
the userid not the username.