Remove authority param from GroupService.create()
This parameter isn't needed because the authority can be retrieved from
the user who is creating the group. Also, it's not actually valid /
doesn't make any sense for the group's authority to be different from
that of the group's creator, but having authority as a separate param
allows this to happen.

Note that one of the tests was actually passing in an authority
different from that of the creator - I had to fix the test.

Also note that you can still create an open group in any authority using
the --authority argument to the add-open-group CLI. The CLI code uses
this authority argument to form the userid of the user's creator, not to
pass it in to group_svc.create().