Py 3 compat - Fix text handling issues in JSON schema tests
Using bytes keys in the test schema and string fields in the test case
JSON input caused validation to fail in Python 3.

For Python 3, use unicode keys in the test schema and JSON input to
resolve the issue. Using Unicode literals for property names in Python 2
results in different exception messages (`u'foo'` vs `'foo'`). Work
around this by using `bytes` as the type for property names in Py 2.