Py 3 compat - Fix charset issue
When overriding the Content-Type params in `read_jsonld`, the charset
associated with the response was lost inside `webob`, causing a later
error when Pyramid tried to assign to `response.text` when rendering the
JSON result.

This didn't happen in Python 2 because `webob.request` executes
different code paths in that environment.

There have been changes in this area in webob >= 1.7, so updating webob
may also be a solution, but that may have other consequences so I felt
this was a smaller and safer change to solve the problem at hand.