Add new GroupScope model class
Add a new GroupScope model class and tests. No code uses this class yet,
though. (A Group can optionally be related to one or more GroupScopes
but no production code relates any Groups to any GroupScopes yet.)

The "scope" of a group is, currently, the list of one or more hostnames
that the group can be used on (when the client is launched on a URL
matching one of those hostnames).

The generic term "scope" is used because we might want to extend a
group's scope beyond just hostnames in the future. E.g. allow scopes to
be other kinds of URIs, or allow a scope to contain both a hostname and
a path.

For now only hostnames are supported, so the `groupscope` table has just
a single `hostname` column.

These "hostnames" are actually just strings - there's no validation yet.

There's a many-many relationship between groups and scopes:

    # All the groups related to a given scope.
    my_scope.groups

    # All of a given group's scopes.
    my_group.scopes

It's not mandatory for a group to have any scopes (`my_group.scopes` can
be `[]`), the intention is that in the future open groups will always
have one or more scopes (enforced at the service level) but private
groups will never have any scopes.

Usage of the new `GroupScope` class:

    my_group = models.Group(...)
    my_scope = models.GroupScope(hostname=u'biopub.org')
    my_scope.groups.append(my_group)
    my_scope.groups  # [<Group: foobar>]
    my_group.scopes  # [<GroupScope biopub.org>]

You can also set a group's scopes directly when you create the group:

    models.Group(..., scopes=[my_scope, my_other_scope])

This pull request also contains a [factory_boy](https://factoryboy.readthedocs.io/en/latest/)
factory for conveniently creating GroupScopes in tests.
`factories.GroupScope()` will return a new GroupScope with a randomly
generated hostname and a randomly chosen number of randomly generated
groups:

    s = factories.GroupScope()  # <GroupScope pacheco.net>
    s.hostname                  # u'pacheco.net'
    s.groups                    # [<Group: test-group-0>, <Group: test-group-1>]

(So this means that creating a GroupScope with factories.GroupScope() also
creates some Groups and adds them to the DB session as well.)

You can pass in your own hostname:

    factories.GroupScope(hostname=u'biopub.org')  # <GroupScope biopub.org>

You can control the number of groups by passing in an int as the `groups`
argument. For example to create a GroupScope with no groups:

    s = factories.GroupScope(groups=0)
    s.groups  # []

You can create a GroupScope for some preexisting groups by passing in a
list of `models.Group` objects as the `groups` argument:

    factories.GroupScope(groups=[my_group, another_group])