Attempt to improve an flakey test fail message
Edit this test so that it should be testing the same things as before
but is should give us a better clue as to why exactly the test failed
when it fails.

tests/h/search/index_test.py::TestBatchIndexer::test_index_indexes_all_annotations_to_es
is occassionally failing with this assertion error:

    E AssertionError:
    Expected call: streaming_bulk(
        <Mock name='mock.conn' id='140410258777232'>,
        <iterable with [<Annotation 3l0CPAzaEei6VmNdOzSvCg>, <Annotation 3mUY-gzaEei6Vl97TJaBfg>]>,
        chunk_size=<ANY>,
        expand_action_callback=<ANY>,
        raise_on_error=False)
    Actual call: streaming_bulk(
        <Mock name='mock.conn' id='140410258777232'>,
        <generator object _log_status at 0x7fb3cf345be0>,
        chunk_size=100,
        expand_action_callback=<bound method BatchIndexer._prepare of <h.search.index.BatchIndexer object at 0x7fb3cf284a10>>,
        raise_on_error=False)

The expected call looks the same as the actual call, assuming that
`<iterable with [<Annotation 3l0CPAzaEei6VmNdOzSvCg>, <Annotation 3mUY-gzaEei6Vl97TJaBfg>]>`
matches `<generator object _log_status at 0x7fb3cf345be0>`, which is
usually does (the test usually passes).

My hunch is that the iterable sometimes fails to match the generator
because the order in which the generator returns the two annotations is
sometimes different.