Tighten up spacing in tags section
Tighten up spacing between tags and between the text and the grey
box around it. This was done so it doesn’t create an optical
illusion where the tags look like they are a different distance
away than the user names are from their section titles. Remove
the margin from the tags container-this will be taken care of
later by specifying the margin from each section title to it’s
content so it can be the same for every section. The corners
were also curved on the tag boxes to create more of a tag-like
effect.