Remove extra group creation in fixtures
Creating extra groups is pretty unnecessary as both the group
search controller class and the base class do not have any
group selection logic for the query so it would be fairly
unlikely to introduce a bug there. If this is something to
be tested it should/could be tested at a lower level. Creating
extra groups in the fixtures each time, creates a performance
hit.