Delimit counts on activity pages with ','
Add , as thousands delimiter in annotation count on activity
pages. The annotation count blends together when it's really
large and it makes it hard to read. This formating was added
as a jinja extension and a corresponding test was also added.
This formats the search result Matching # annotations, # of
times a tag was used, number of annoations a particular user
has made, and the group and user side bar total annotations
to be , delimited.