Add day to date on activity page sidebar
Add day to group created date on sidebar so date format is now:
Month day, year aka: August 1, 2018. Some tests were updated to
expect this new date format. Note the date format was changed from
using the c library strftime to using the python format method as
the c library does not support %-d on all systems (a way of removing
the default zero padding of the day). This date format function
was added to the utils area and corresponding tests were added
there as well.