Move group info & links into sidebar section
Previously the group description, number of annotations, date created,
edit group link, and join group link were outside of a sidebar section.
This change moves them into a sidebar section which makes them
easier to format. Since these are already grouped into categories of
information, a div was added around each of them called a ‘subsection’.
This div enables equal spacing between each of the subsections. Since
the dl that was being used is complex to style with this new tag
heirarchy, it was changed into a series of paragraph and span tags.
Note css classes with the prefix subsection were added for styling.