Accept string key material in h.security.derive_key
The secret key is specified via an environment variable so is read into
the app as a string. The `hkdf.derive` method requires it to be `bytes`
in Python 3. Handle this by simply encoding the key material to UTF-8
bytes if needed.

The encoding could have been done at a higher level, eg. when reading
the settings, but this was a little more convenient.