Don't attach throw-away test DocumentURI to DB
Fix an intermittent `IntegrityError` created by `factories.Annotation()`
attaching conflicting `DocumentURI` objects to the test DB session:

The `factories.Annotation()` test factory creates a throw-away
`DocumentURI` object that isn't intended to be attached to the DB
session (it's created with `factories.DocumentURI.build()`, rather than
`factories.DocumentURI()`).

The `Annotation` factory just wants to use the `DocumentURI` factory for
its random autogeneration of `DocumentURI` fields, but doesn't actually
want to keep the created `DocumentURI` object or add it to the DB, it
just uses those field values in a dict that it passes to
`update_document_metadata()` further down.

`update_document_metadata()` will end up creating _another_ new
`DocumentURI` based on the values of this dict or, if a conflicting
`DocumentURI` already exists in the DB, it will just reuse that one
instead of creating a new one. This reuse behaviour prevents any
duplicate `DocumentURI`s getting added to the DB and causing
`IntegrityError`s.

The **problem** is that the `Annotation` factory was attaching the
temporary, throw-away `DocumentURI` URI to its own `self.document`:

    document_uri = DocumentURI.build(document=self.document,

`self.document` is an ORM object that _is_ attached to the DB session.
The `DocumentURI` created using `.build()` would not normally be
attached to the DB session but, since it's now connected to another ORM
object that is attached, it gets indirectly attached to the DB session.
Oops!

Furthermore, there appears to be no reason why the throw-away
`DocumentURI` needs to be attached to `self.document`.

Since there's nothing to prevent the throw-away `DocumentURI` being a
duplicate of an already-existing `DocumentURI`, and
`factories.DocumentURI`'s random generation will sometimes produce
duplicates, this sometimes causes an `IntegrityError`.

Fix this by simply not attaching the `DocumentURI` to `self.document`.
Explicitly pass `document=None` because otherwise
`factories.DocumentURI` would generate a new `Document` for the
`DocumentURI` to be attached to and we don't want that either.